refactor: svg icon 일부 public -> src로 디렉토리 변경
